projects/rag-stuff/pyproject.toml
[project]
name = "rag-stuff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "quotientai",
    "litellm",
    "requests",
    "python-dotenv",
    "cohere>=5.14.0",
]
